First cut at the Vision Tools.
This is a tool that will allow us to edit each part of the vision
algorithm. Currently the following steps are implemented:
1. Threshold Color (on the HSV channels)
2. Canny Transformation
3. Hough Transformation.

To add a new step, implement the ITransformation interface (You can also
extend the AbstractTransformation) and register it with the
VisionToolsControl class. All of the registration is done in the
VisionToolsMain class. The main requirement for the applyTransformation
method is that the input and output image must be of type BGR
(BufferedImage Type TYPE_3BYTE_BGR). Eventaully I want to change this so
that both the image and artifacts (such as points) are added.